requirements.txt
accelerate==1.12.0
aiohappyeyeballs==2.6.1
aiohttp==3.13.2
aiosignal==1.4.0
annotated-types==0.7.0
anyio==4.12.0
attrs==25.4.0
audioop-lts==0.2.2
beautifulsoup4==4.14.3
blinker==1.9.0
cachetools==6.2.4
certifi==2025.11.12
charset-normalizer==3.4.4
click==8.3.1
colorama==0.4.6
diffusers==0.36.0
discord.py==2.6.4
distro==1.9.0
duckduckgo_search==8.1.1
ffmpeg==1.4
ffprobe==0.5
filelock==3.20.1
Flask==3.1.2
frozenlist==1.8.0
fsspec==2025.12.0
google-api-core==2.28.1
google-api-python-client==2.187.0
google-auth==2.45.0
google-auth-httplib2==0.3.0
googleapis-common-protos==1.72.0
googlesearch-python==1.3.0
groq==1.0.0
h11==0.16.0
hf-xet==1.2.0
httpcore==1.0.9
httplib2==0.31.0
httpx==0.28.1
huggingface-hub==0.36.0
idna==3.11
importlib_metadata==8.7.0
itsdangerous==2.2.0
Jinja2==3.1.6
lxml==6.0.2
MarkupSafe==3.0.3
mpmath==1.3.0
multidict==6.7.0
networkx==3.6.1
numpy==2.3.5
ollama==0.6.1
packaging==25.0
pillow==12.0.0
primp==0.15.0
propcache==0.4.1
proto-plus==1.27.0
protobuf==6.33.2
psutil==7.1.3
pyasn1==0.6.1
pyasn1_modules==0.4.2
pydantic==2.12.5
pydantic_core==2.41.5
pydub==0.25.1
pyparsing==3.2.5
python-dotenv==1.2.1
PyYAML==6.0.3
regex==2025.11.3
requests==2.32.5
rsa==4.9.1
safetensors==0.7.0
setuptools==80.9.0
shellingham==1.5.4
sniffio==1.3.1
soupsieve==2.8.1
sympy==1.14.0
tokenizers==0.22.1
torch==2.9.0+cu130
torchaudio==2.9.0+cu130
torchvision==0.24.0+cu130
tqdm==4.67.1
transformers==4.57.3
typer-slim==0.20.0
typing-inspection==0.4.2
typing_extensions==4.15.0
uritemplate==4.2.0
urllib3==2.6.2
waitress==3.0.2
Werkzeug==3.1.4
yarl==1.22.0
yt-dlp==2025.12.8
zipp==3.23.0
